--- app/javascript/components/widgets/land-cover/fao-cover/settings.js
@@ -1,3 +1,0 @@
-export default {
-  unit: 'ha'
-};
