--- app/javascript/components/widgets/land-cover/global-land-cover/actions.js
@@ -1,7 +1,0 @@
-import { getGlobalLandCover } from 'services/forest-data';
-
-export default ({ params }) =>
-  getGlobalLandCover(params).then(response => {
-    const data = response.data.rows;
-    return data;
-  });
