+++ app/javascript/components/widgets/land-cover/global-land-cover/index.js
@@ -1,7 +1,39 @@
-import Component from 'components/widgets/components/widget-pie-chart-legend';
-import getData from './actions';
-import getProps from './selectors';
-import config from './config';
-import settings from './settings';
+import { getGlobalLandCover } from 'services/forest-data';
 
-export { getData, getProps, Component, config, settings };
+import getWidgetProps from './selectors';
+
+export default {
+  widget: 'globalLandCover',
+  title: 'Land cover for {location}',
+  categories: ['land-cover'],
+  types: ['country'],
+  admins: ['adm0', 'adm1'],
+  options: {
+    years: [2015]
+  },
+  chartType: 'pieChart',
+  analysis: true,
+  datasets: [
+    {
+      dataset: '588f2f1f-cc62-46aa-9859-befa031412ca',
+      layers: ['c09767f5-0ff0-419b-a21b-1b0b06f4745f']
+    }
+  ],
+  colors: 'plantations',
+  metaKey: 'widget_land_cover_esa',
+  hideSettings: true,
+  sortOrder: {
+    landCover: 100
+  },
+  settings: {
+    year: 2015
+  },
+  sentences:
+    'The land use of {location} in {year} is mostly {category}, covering an area of {extent}.',
+  getData: params =>
+    getGlobalLandCover(params).then(response => {
+      const data = response.data.rows;
+      return data;
+    }),
+  getWidgetProps
+};
