+++ app/javascript/components/widgets/land-cover/global-land-cover/selectors.js
@@ -11,7 +11,7 @@ const getData = state => state.data;
 const getSettings = state => state.settings;
 const getLocationName = state => state.locationName;
 const getColors = state => state.colors;
-const getSentences = state => state.config && state.config.sentences;
+const getSentences = state => state.sentences;
 
 // get lists selected
 export const parseData = createSelector(
