--- app/javascript/components/widgets/land-cover/global-land-cover/settings.js
@@ -1,3 +1,0 @@
-export default {
-  year: 2015
-};
