--- app/javascript/components/widgets/land-cover/intact-tree-cover/settings.js
@@ -1,6 +1,0 @@
-export default {
-  forestType: 'ifl',
-  threshold: 30,
-  extentYear: 2010,
-  ifl: 2016
-};
