+++ app/javascript/components/widgets/land-cover/primary-forest/selectors.js
@@ -6,9 +6,9 @@ import { format } from 'd3-format';
 const getData = state => state.data;
 const getSettings = state => state.settings;
 const getLocationName = state => state.locationName;
-const getIndicator = state => state.indicator || null;
+const getIndicator = state => state.indicator;
 const getColors = state => state.colors;
-const getSentences = state => state.config && state.config.sentences;
+const getSentences = state => state.sentences;
 
 // get lists selected
 export const parseData = createSelector(
