--- app/javascript/components/widgets/land-cover/primary-forest/settings.js
@@ -1,5 +1,0 @@
-export default {
-  forestType: 'primary_forest',
-  threshold: 30,
-  extentYear: 2000
-};
