--- app/javascript/components/widgets/land-cover/ranked-plantations/actions.js
@@ -1,28 +1,0 @@
-import {
-  getExtentGrouped,
-  getAreaIntersectionGrouped
-} from 'services/forest-data';
-import axios from 'axios';
-
-export default ({ params }) =>
-  axios
-    .all([
-      getExtentGrouped(params),
-      getAreaIntersectionGrouped({ ...params, forestType: 'plantations' })
-    ])
-    .then(
-      axios.spread((extentGrouped, plantationsExtentResponse) => {
-        let data = {};
-        const extent = extentGrouped.data && extentGrouped.data.data;
-        const plantationsExtent =
-          plantationsExtentResponse.data && plantationsExtentResponse.data.data;
-        if (extent.length && plantationsExtent.length) {
-          data = {
-            extent,
-            plantations: plantationsExtent
-          };
-        }
-
-        return data;
-      })
-    );
