--- app/javascript/components/widgets/land-cover/ranked-plantations/settings.js
@@ -1,6 +1,0 @@
-export default {
-  threshold: 10,
-  pageSize: 5,
-  page: 0,
-  extentYear: 2000
-};
