--- app/javascript/components/widgets/land-cover/tree-cover-located/settings.js
@@ -1,8 +1,0 @@
-export default {
-  threshold: 30,
-  extentYear: 2010,
-  unit: '%',
-  pageSize: 5,
-  page: 0,
-  ifl: 2000
-};
