--- app/javascript/components/widgets/land-cover/tree-cover-plantations/config.js
@@ -1,31 +1,0 @@
-export default {
-  widget: 'treeCoverPlantations',
-  title: 'Plantations in {location}',
-  categories: ['land-cover'],
-  types: ['country'],
-  admins: ['adm0', 'adm1', 'adm2'],
-  colors: 'plantations',
-  datasets: [
-    {
-      // global plantations
-      dataset: 'bb1dced4-3ae8-4908-9f36-6514ae69713f',
-      layers: ['b8fb6cc8-6893-4ae0-8499-1ca9f1ababf4']
-    }
-  ],
-  analysis: true,
-  metaKey: 'widget_plantation_extent',
-  sortOrder: {
-    landCover: 100
-  },
-  sentences: {
-    initialSpecies:
-      'In {location}, {firstSpecies} and {secondSpecies} represent the largest plantation area by {type}, spanning {extent} and {percent} of land area.',
-    singleSpecies:
-      'In {location}, {firstSpecies} represent the largest plantation area by {type}, spanning {extent} and {percent} of land area.',
-    initialTypes:
-      'In {location}, the largest plantation area by type is {topType}, spanning {extent} and {percent} of land area.'
-  },
-  whitelists: {
-    indicators: ['plantations']
-  }
-};
