--- app/javascript/components/widgets/land-cover/tree-cover-plantations/settings.js
@@ -1,5 +1,0 @@
-export default {
-  threshold: 10,
-  type: 'bound2',
-  extentYear: 2010
-};
