--- app/javascript/components/widgets/land-cover/tree-cover-ranked/settings.js
@@ -1,6 +1,0 @@
-export default {
-  threshold: 30,
-  unit: '%',
-  extentYear: 2000,
-  ifl: 2000
-};
